Split compare scatter tests into one class per MVC component
